feat: ft-wave2-factory-catalog-discovery-004 - Close ownership metadata and verification gates
